CommandResult: encapsulated feedbackToUser
The member `feedbackToUser` is currently exposed to members outside
of the class

Without encapsulation, we cannot prevent illegal direct access of its
member

Let's us encapsulate and hide its data and add a new getter method
which is exposed to the public. This will provide us with a better
flexibility and easier maintainability in the future.