CommandResult class: encapsulate member `feedbackToUser`
The class variable `feedbackToUser` of `CommandResult` class is set to
`public` and `final`, which allows the variable to be visible but
not modifiable from outside of a `CommandResult` object.

If `feedbackToUser` is no longer set to `final` in the future, the
variable will become mutable and modifiable from outside of a
`CommandResult` object.

This allows unintended actions such as inadvertent modification to the
`feedbackToUser` variable from outside of a `CommandResult` object.

To future-proof the `CommandResult` class, let's err on the side of
caution and encapsulate `feedbackToUser` so that it can only be accessed
using a getter method in the `CommandResult` class.