Name: implements `Comparable` interface
We are going to add a new `sort` command to sort all persons in the
address book by their names lexicographically.

As a first step, let's make the Name class implement the `Comparable`
interface so that it becomes possible to compare Name objects and allow
for sorting in the future.