commands: implement `sort` command
We are ready to implement the `sort` command.

The `sort` command helps to sort all persons in the address book by
their names lexicographically. The command should be handled by the
parser, and display the sample command usage in the `help` command so as
to prevent confusion among users.

Let's,
* add a SortCommand class that calls AddressBook#sort
* handle the SortCommand#COMMAND_WORD in Parser class
* update the command usage in HelpCommand class